+++ src/controllers/service-accounts/list-service-accounts.ts
@@ -1,23 +1,11 @@
-import {AppRouteHandler, Response} from '@gravity-ui/expresskit';
+import {AppRouteHandler} from '@gravity-ui/expresskit';
 
 import {ApiTag} from '../../components/api-docs';
-import {z} from '../../components/zod';
 import {CONTENT_TYPE_JSON} from '../../constants/content-type';
 import {listServiceAccounts} from '../../services/service-accounts/list-service-accounts';
 import {serviceAccountModelArray} from '../reponse-models/service-accounts/service-account-model-array';
 
-const responseSchema = z
-    .object({
-        serviceAccounts: serviceAccountModelArray.schema,
-    })
-    .describe('Service accounts list');
-
-type ResponseBody = z.infer<typeof responseSchema>;
-
-export const listServiceAccountsController: AppRouteHandler = async (
-    req,
-    res: Response<ResponseBody>,
-) => {
+export const listServiceAccountsController: AppRouteHandler = async (req, res) => {
     const result = await listServiceAccounts({ctx: req.ctx});
 
     res.status(200).send({
@@ -30,10 +18,10 @@ listServiceAccountsController.api = {
     tags: [ApiTag.Management],
     responses: {
         200: {
-            description: responseSchema.description ?? '',
+            description: serviceAccountModelArray.schema.description ?? '',
             content: {
                 [CONTENT_TYPE_JSON]: {
-                    schema: responseSchema,
+                    schema: serviceAccountModelArray.schema,
                 },
             },
         },
